+++ __tests__/extended_asset.test.ts
@@ -1,0 +1,18 @@
+// import { asset, symbol } from "..";
+
+// const asset_mask: bigint = (BigInt(1) << BigInt(62)) - BigInt(1);
+// const asset_min: bigint = -asset_mask; // -4611686018427387903
+// const asset_max: bigint = asset_mask; //  4611686018427387903
+
+test("extended_asset_type_test", () => {
+    // const sym_no_prec = symbol("SYMBOLL", 0);
+    // const sym_prec = symbol("SYMBOLL", 63);
+
+    // const ext_sym_no_prec = extended_sybmol( sym_no_prec, name("eosioaccountj"));
+    // const ext_sym_prec = extended_sybmol( sym_prec, name("eosioaccountj"));
+
+    // asset_no_prec = asset(0, sym_no_prec );
+    // asset_prec = asset(0, sym_prec );
+
+    expect(true).toBeTruthy();
+});